+++ py_trees_ros/programs/blackboard_watcher.py
@@ -27,9 +27,10 @@ import functools
 import py_trees_msgs.srv as py_trees_srvs
 import rospy
 import py_trees.console as console
-import rosservice
 import sys
 import std_msgs.msg as std_msgs
+
+from . import utilities
 
 ##############################################################################
 # Classes
@@ -123,7 +124,7 @@ def watch_sub_blackboard(received_topic, namespace):
     """
     rospy.init_node(received_topic.split('/')[-1])
 
-    close_blackboard_watcher_service_name = find_service(namespace, 'py_trees_msgs/CloseBlackboardWatcher')
+    close_blackboard_watcher_service_name = utilities.find_service(namespace, 'py_trees_msgs/CloseBlackboardWatcher')
 
     def request_close_blackboard_watcher(updates_subscriber):
         """
@@ -148,58 +149,11 @@ def watch_sub_blackboard(received_topic, namespace):
     while not rospy.is_shutdown():
         rospy.spin()
 
-def discover_namespace(suggested_namespace):
-    try:
-        service_name_list = rosservice.rosservice_find("py_trees_msgs/OpenBlackboardWatcher")
-    except rosservice.ROSServiceIOException as e:
-        print(console.red + "ERROR: {0}".format(str(e)) + console.reset)
-        sys.exit(1)
-
-    if not service_name_list:
-        print(console.red + "ERROR: blackboard services not found" + console.reset)
-        sys.exit(1)
-
-    if suggested_namespace is not None:
-        for service_name in service_name_list:
-            if suggested_namespace in service_name:
-                return suggested_namespace
-        print("")
-        print(console.red
-              + "ERROR: blackboard services found {}".format(service_name_list)
-              + console.reset
-        )
-        print(console.red
-              + "ERROR: but none matching the requested '{}'".format(suggested_namespace)
-              + console.reset)
-        print("")
-        sys.exit(1)
-
-    if len(service_name_list) > 1:
-        print(console.red + "\nERROR: multiple blackboard services found %s" % service_name_list + console.reset)
-        print(console.red + "\nERROR: select one with the --namespace argument" + console.reset)
-        sys.exit(1)
-
-    # the previous checks guarantee we have a single name here
-    service_name = service_name_list[0]
-    # drop the last substring
-    namespace = '/'.join(service_name.split('/')[:-1])
-    return namespace
-        
-def find_service(namespace, service_type):
-    # This method assumes all error handling for discovery of a unique
-    # services associated with the namespace, service_type pair has been
-    # hitherto handled by a call to discover_namespace
-    try:
-        service_name = rosservice.rosservice_find(service_type)
-    except rosservice.ROSServiceIOException as e:
-        print(console.red + "ERROR: {0}".format(str(e)) + console.reset)
-        sys.exit(1)
-    return service_name[0]
 
 def handle_args(args):
-    args.namespace = discover_namespace(args.namespace)
+    args.namespace = utilities.discover_namespace(args.namespace)
     if args.list_variables:
-        list_variables_service_name = find_service(args.namespace, 'py_trees_msgs/GetBlackboardVariables')
+        list_variables_service_name = utilities.find_service(args.namespace, 'py_trees_msgs/GetBlackboardVariables')
         try:
             rospy.wait_for_service(list_variables_service_name, timeout=3.0)
             try:
@@ -219,7 +173,7 @@ def handle_args(args):
             variables = args.variables[0:]
             variables = [variable.strip(',[]') for variable in variables]
 
-            open_blackboard_watcher_service = find_service(args.namespace, 'py_trees_msgs/OpenBlackboardWatcher')
+            open_blackboard_watcher_service = utilities.find_service(args.namespace, 'py_trees_msgs/OpenBlackboardWatcher')
 
             try:
                 rospy.wait_for_service(open_blackboard_watcher_service, timeout=3.0)